Remove function calls for unsupported KDFs
