+++ app/implementations/openssl/3/iut.h
@@ -49,12 +49,8 @@ int app_sha_handler(ACVP_TEST_CASE *test_case);
 int app_hmac_handler(ACVP_TEST_CASE *test_case);
 int app_cmac_handler(ACVP_TEST_CASE *test_case);
 int app_kmac_handler(ACVP_TEST_CASE *test_case);
-int app_kdf135_snmp_handler(ACVP_TEST_CASE *test_case);
 int app_kdf135_ssh_handler(ACVP_TEST_CASE *test_case);
-int app_kdf135_srtp_handler(ACVP_TEST_CASE *test_case);
-int app_kdf135_ikev2_handler(ACVP_TEST_CASE *test_case);
 int app_kdf108_handler(ACVP_TEST_CASE *test_case);
-int app_kdf135_ikev1_handler(ACVP_TEST_CASE *test_case);
 int app_kdf135_x942_handler(ACVP_TEST_CASE *test_case);
 int app_kdf135_x963_handler(ACVP_TEST_CASE *test_case);
 int app_pbkdf_handler(ACVP_TEST_CASE *test_case);
